requirements.txt
# pip install -r requirements.txt

# Основные зависимости для Telegram-бота (с поддержкой JobQueue)
python-telegram-bot[job-queue]==20.7

# Работа с переменными окружения (.env)
python-dotenv==1.0.1

# База данных (SQLite через SQLAlchemy, можно будет легко перейти на PostgreSQL/MySQL)
SQLAlchemy==2.0.23

# Работа с датами и временем
python-dateutil==2.9.0.post0

# Для HTTP-запросов (например, API погоды, праздники, OpenAI)
requests==2.31.0

# OpenAI API (ChatGPT интеграция в будущем)
openai==1.3.7

# Для удобства при будущей работе с JSON/YAML конфигами
pyyaml==6.0.1
